--- src/constants/sdKECommerce.ts
@@ -1,6 +1,0 @@
-export const CTP_PROJECT_KEY = 'bon747jour';
-export const CTP_CLIENT_SECRET = 'FdHmxPyT12o4qb0VWg4z5k_glAZI5vbt';
-export const CTP_CLIENT_ID = 'SiyGK3A8uQpMq5CgHb3egU9K';
-export const CTP_AUTH_URL = 'https://auth.europe-west1.gcp.commercetools.com';
-export const CTP_API_URL = 'https://api.europe-west1.gcp.commercetools.com';
-export const CTP_SCOPES = 'manage_project:bon747jour';
